9/12/2014 11:28am Save Point (for real)
Ok, now I'm good.

+++ Square9Analytics.API/Controllers/ActionsController.cs
@@ -11,7 +11,7 @@ namespace Square9Analytics.Controllers
 {
     /// <summary>
     /// API Endpoint for action-based analytics data requests.
-    /// 9/12/2014 11:08 am save point
+    /// 9/12/2014 11:28 am save point
     /// </summary>
     public class ActionsController : AnalyticsController
     {
@@ -34,7 +34,7 @@ namespace Square9Analytics.Controllers
                 //Validates the startdate and endDate strings as a dates
                 if (DateTime.TryParse(startdate, out StartdateValue) && DateTime.TryParse(endDate, out EnddateValue))
                 {
-                    Int32 docCount = getNumOfDocs.getActionCount(StartdateValue, EnddateValue, AuditEntry.DocumentIndexed); //need the object to pass startdate and enddate to
+                    float docCount = getNumOfDocs.getActionCount(StartdateValue, EnddateValue, AuditEntry.DocumentIndexed); //need the object to pass startdate and enddate to
 
                     return Request.CreateResponse(HttpStatusCode.OK, docCount);
                 }
@@ -61,7 +61,7 @@ namespace Square9Analytics.Controllers
                 //Validates the startdate and endDate strings as a dates
                 if (DateTime.TryParse(startdate, out StartdateValue) && DateTime.TryParse(endDate, out EnddateValue))
                 {
-                    Int32 workflowCount = getNumOfWorkflows.getActionCount(StartdateValue, EnddateValue, AuditEntry.DocumentIndexed); //need the object to pass startdate and enddate to
+                    float workflowCount = getNumOfWorkflows.getActionCount(StartdateValue, EnddateValue, AuditEntry.DocumentIndexed); //need the object to pass startdate and enddate to
 
                     return Request.CreateResponse(HttpStatusCode.OK, workflowCount);
                 }
